refactor: consolidate platform typing into single torch-based module.
- move Platform/PlatformEnum/CpuArchEnum/detect_hardware into
  xllm/python/platform.py as the single home; auto_config loads it by
  path and re-exports for the torchless launcher
- adopt sglang-style current_platform singleton; rename is_gpu to is_cuda
- detect device type solely from the framework runtime, dropping the
  visible-device env-mask and npu-smi fallbacks
- drop unused ABC base from BaseTuner
