+++ xllm/auto_config/utils.py
@@ -13,357 +13,64 @@
 # limitations under the License.
 # ==============================================================================
 
-"""Hardware typing for xLLM auto-tuning.
+"""Auto-tuning machinery for `xllm serve`.
 
-A slimmed-down Python `Platform`, modeled on sglang's
-`multimodal_gen/runtime/platforms/interface.py`, that answers "what hardware is
-this?" for the auto_config tuning profiles. It runs inside the `xllm serve`
-launcher process, which has no hard `torch` dependency, so detection is
-layered and torch-optional:
+Hosts the per-model tuning base class (`BaseTuner`) and the launcher helpers
+that resolve a model's tuning profile and generate its tuned config.
 
-1. Prefer the framework runtime (`torch` / `torch_npu` / `torch_mlu` / ...)
-   when it is importable and reports an available device -- the most accurate
-   signal, matching `scripts/build_support/utils.py::get_device_type`.
-2. Otherwise fall back to visible-device env masks, then `npu-smi` for the
-   Ascend chip name, neither of which requires torch.
-
-Nothing here raises: an undetectable environment resolves to
-`PlatformEnum.UNSPECIFIED` / `CpuArchEnum.UNSPECIFIED` / `"unknown"` so callers
-can always read a value.
+Hardware typing (`Platform`, `PlatformEnum`, `CpuArchEnum`, `detect_hardware`)
+lives in `xllm/python/platform.py`, the single home shared with the C++ worker.
+The launcher has no hard `torch` dependency and cannot `import
+xllm.python.platform` normally -- that would execute `xllm/python/__init__.py`,
+which imports `torch` and binds a kernel package. So this module loads that one
+file by path (the same technique `load_tuning_module` uses for tuning profiles)
+and re-exports the symbols, keeping the tuning profiles' imports unchanged.
 """
 
 from __future__ import annotations
 
 import copy
-import enum
-import functools
 import importlib.util
 import json
 import os
-import platform as platform_module
-import re
-import subprocess
-from abc import ABC, abstractmethod
 from types import ModuleType
 from typing import Any, Dict, Optional, Sequence
 
 from scripts.logger import logger
 
 
-class PlatformEnum(enum.Enum):
-    CUDA = enum.auto()
-    NPU = enum.auto()
-    MLU = enum.auto()
-    MUSA = enum.auto()
-    DCU = enum.auto()
-    ILU = enum.auto()
-    CPU = enum.auto()
-    UNSPECIFIED = enum.auto()
+def _load_platform_module() -> ModuleType:
+    """Load `xllm/python/platform.py` by file path, dodging its package init.
 
-
-class CpuArchEnum(enum.Enum):
-    X86 = enum.auto()
-    ARM = enum.auto()
-    UNSPECIFIED = enum.auto()
-
-
-# Visible-device env masks per backend, mirroring the CLI reference's
-# device-selection note. Order matches the accelerator detection precedence.
-_VISIBLE_DEVICE_ENV_VARS = {
-    PlatformEnum.NPU: "ASCEND_RT_VISIBLE_DEVICES",
-    PlatformEnum.CUDA: "CUDA_VISIBLE_DEVICES",
-    PlatformEnum.MLU: "MLU_VISIBLE_DEVICES",
-    PlatformEnum.DCU: "HIP_VISIBLE_DEVICES",
-    PlatformEnum.MUSA: "MUSA_VISIBLE_DEVICES",
-}
-
-
-def _torch_device_type() -> Optional[PlatformEnum]:
-    """Detect the accelerator via the framework runtime, or None if absent.
-
-    Follows the same probing order as
-    `scripts/build_support/utils.py::get_device_type`. Every import is guarded:
-    the launcher must not fail just because a framework wheel is missing.
+    `platform.py` only imports the stdlib and the shared logger, so it loads
+    cleanly here without pulling in `torch` or a kernel package. A private
+    module name keeps it from clashing with the worker's real
+    `xllm.python.platform`.
     """
-    try:
-        import torch
-    except ImportError:
-        return None
-
-    try:
-        if torch.cuda.is_available():
-            try:
-                from torch.utils.cpp_extension import HIP_HOME
-
-                if HIP_HOME and "dtk" in HIP_HOME.lower():
-                    return PlatformEnum.DCU
-            except ImportError:
-                pass
-            try:
-                import ixformer  # noqa: F401
-
-                return PlatformEnum.ILU
-            except ImportError:
-                return PlatformEnum.CUDA
-    except Exception:
-        pass
-
-    for module_name, attr, enum_value in (
-        ("torch_musa", "musa", PlatformEnum.MUSA),
-        ("torch_mlu", "mlu", PlatformEnum.MLU),
-        ("torch_npu", "npu", PlatformEnum.NPU),
-    ):
-        try:
-            __import__(module_name)
-            device_module = getattr(torch, attr, None)
-            if device_module is not None and device_module.is_available():
-                return enum_value
-        except ImportError:
-            continue
-        except Exception:
-            continue
-
-    return None
+    platform_path = os.path.join(
+        os.path.dirname(os.path.dirname(os.path.realpath(__file__))),
+        "python",
+        "platform.py",
+    )
+    spec = importlib.util.spec_from_file_location(
+        "xllm.auto_config._platform", platform_path
+    )
+    if spec is None or spec.loader is None:
+        raise ImportError(f"failed to load platform module: {platform_path}")
+    module = importlib.util.module_from_spec(spec)
+    spec.loader.exec_module(module)
+    return module
 
 
-def _env_device_type() -> Optional[PlatformEnum]:
-    """Detect the accelerator from visible-device env masks, or None."""
-    for enum_value, env_var in _VISIBLE_DEVICE_ENV_VARS.items():
-        if os.environ.get(env_var) is not None:
-            return enum_value
-    return None
+_platform_module = _load_platform_module()
 
-
-def _npu_chip_from_smi() -> Optional[str]:
-    """Best-effort lowercase Ascend chip name from `npu-smi info`, or None.
-
-    `npu-smi` may be absent or permission-blocked, so this never raises. The
-    chip appears in the per-device rows (e.g. `910B2C`).
-    """
-    try:
-        completed = subprocess.run(
-            ["npu-smi", "info"],
-            capture_output=True,
-            text=True,
-            timeout=15,
-            check=False,
-        )
-    except (OSError, subprocess.SubprocessError):
-        return None
-
-    match = re.search(r"910[A-Za-z0-9]*", completed.stdout)
-    if match is None:
-        return None
-    return match.group(0).lower()
-
-
-def _npu_chip_from_acl() -> Optional[str]:
-    """Best-effort lowercase Ascend chip name via `acl`, or None.
-
-    `acl` is often unavailable in the launcher's import path and may already be
-    initialized by torch_npu, so failures are swallowed and this is only a
-    supplement to `npu-smi`.
-    """
-    try:
-        import acl
-    except ImportError:
-        return None
-    try:
-        soc_name = acl.get_soc_name()
-    except Exception:
-        return None
-    if not soc_name:
-        return None
-    match = re.search(r"910[A-Za-z0-9]*", soc_name)
-    if match is None:
-        return None
-    return match.group(0).lower()
-
-
-class Platform:
-    """Streamlined hardware-typing facade for auto-tuning profiles.
-
-    All methods are classmethods so profiles can call `Platform.is_npu()`
-    without holding an instance. Detection results are cached for the process
-    lifetime.
-    """
-
-    @classmethod
-    @functools.lru_cache(maxsize=1)
-    def enum(cls) -> PlatformEnum:
-        detected = _torch_device_type()
-        if detected is not None:
-            return detected
-
-        detected = _env_device_type()
-        if detected is not None:
-            return detected
-
-        # No framework runtime and no visible-device mask, but a readable
-        # Ascend chip still means this is an NPU host.
-        if cls.get_npu_chip() != "unknown":
-            return PlatformEnum.NPU
-
-        logger.warning(
-            "Platform: could not detect an accelerator; treating host as CPU."
-        )
-        return PlatformEnum.CPU
-
-    @classmethod
-    def is_cuda(cls) -> bool:
-        return cls.enum() == PlatformEnum.CUDA
-
-    @classmethod
-    def is_npu(cls) -> bool:
-        return cls.enum() == PlatformEnum.NPU
-
-    @classmethod
-    def is_mlu(cls) -> bool:
-        return cls.enum() == PlatformEnum.MLU
-
-    @classmethod
-    def is_musa(cls) -> bool:
-        return cls.enum() == PlatformEnum.MUSA
-
-    @classmethod
-    def is_dcu(cls) -> bool:
-        return cls.enum() == PlatformEnum.DCU
-
-    @classmethod
-    def is_ilu(cls) -> bool:
-        return cls.enum() == PlatformEnum.ILU
-
-    @classmethod
-    def is_cpu(cls) -> bool:
-        return cls.enum() == PlatformEnum.CPU
-
-    @classmethod
-    def device_type(cls) -> str:
-        """Lowercase device-type string, aligned with build-time detection."""
-        return cls.enum().name.lower()
-
-    @classmethod
-    @functools.lru_cache(maxsize=1)
-    def get_cpu_architecture(cls) -> CpuArchEnum:
-        arch = platform_module.machine().lower()
-        if "x86" in arch or "amd64" in arch:
-            return CpuArchEnum.X86
-        if "arm" in arch or "aarch64" in arch:
-            return CpuArchEnum.ARM
-        return CpuArchEnum.UNSPECIFIED
-
-    @classmethod
-    def get_cpu_arch_str(cls) -> str:
-        """Raw CPU machine string, e.g. `x86_64` or `aarch64`."""
-        return platform_module.machine() or "unknown"
-
-    @classmethod
-    @functools.lru_cache(maxsize=1)
-    def get_npu_chip(cls) -> str:
-        """Lowercase NPU chip identifier (e.g. `910b2c`), or `"unknown"`.
-
-        Prefers `npu-smi` (reliable in the launcher) and falls back to `acl`.
-        """
-        chip = _npu_chip_from_smi()
-        if chip is not None:
-            return chip
-        chip = _npu_chip_from_acl()
-        if chip is not None:
-            return chip
-        return "unknown"
-
-    @classmethod
-    @functools.lru_cache(maxsize=1)
-    def get_ascend_soc_generation(cls) -> Optional[str]:
-        """Ascend SoC generation (`a2`/`a3`/`a5`), or None if not an NPU host.
-
-        Derived from the chip name so it does not require a second `acl.init`.
-        Mapping matches `scripts/build_support/utils.py::get_ascend_platform`:
-        910B -> a2, 910C / 910_93 -> a3, 950 -> a5, other 910 -> a2.
-        """
-        chip = cls.get_npu_chip()
-        if chip == "unknown":
-            return None
-        if chip.startswith("910b"):
-            return "a2"
-        if chip.startswith("910c") or "910_93" in chip:
-            return "a3"
-        if chip.startswith("950"):
-            return "a5"
-        if chip.startswith("910"):
-            return "a2"
-        return None
-
-    @classmethod
-    def get_device_count(cls) -> Optional[int]:
-        """Number of usable devices, or None if undetectable.
-
-        Prefers the framework runtime's device count; falls back to counting
-        entries in the platform's visible-device env mask.
-        """
-        try:
-            import torch
-        except ImportError:
-            torch = None
-
-        if torch is not None:
-            try:
-                if cls.is_cuda() or cls.is_dcu() or cls.is_ilu():
-                    return torch.cuda.device_count()
-                device_module = getattr(torch, cls.device_type(), None)
-                if device_module is not None and hasattr(
-                    device_module, "device_count"
-                ):
-                    return device_module.device_count()
-            except Exception:
-                pass
-
-        env_var = _VISIBLE_DEVICE_ENV_VARS.get(cls.enum())
-        if env_var is not None:
-            value = os.environ.get(env_var)
-            if value is not None:
-                entries = [
-                    entry for entry in value.split(",") if entry.strip() != ""
-                ]
-                return len(entries)
-        return None
-
-    @classmethod
-    def get_device_name(cls, device_id: int = 0) -> str:
-        """Human-readable device name, best-effort.
-
-        Uses the framework runtime when available; otherwise composes a name
-        from the device type and (for NPU) the chip.
-        """
-        try:
-            import torch
-
-            if cls.is_cuda() or cls.is_dcu() or cls.is_ilu():
-                return torch.cuda.get_device_name(device_id)
-            device_module = getattr(torch, cls.device_type(), None)
-            if device_module is not None and hasattr(
-                device_module, "get_device_name"
-            ):
-                return device_module.get_device_name(device_id)
-        except Exception:
-            pass
-
-        if cls.is_npu():
-            return f"npu:{cls.get_npu_chip()}"
-        return cls.device_type()
-
-
-def detect_hardware() -> Dict[str, str]:
-    """Return the current hardware descriptor: CPU arch, device type, NPU chip.
-
-    Backed by `Platform` so detection stays consistent across profiles. Every
-    field falls back gracefully so the caller can always rely on the keys.
-    """
-    return {
-        "arch": Platform.get_cpu_arch_str(),
-        "device_type": Platform.device_type(),
-        "chip": Platform.get_npu_chip(),
-    }
+# Re-export the shared hardware-typing surface so tuning profiles keep importing
+# it from `xllm.auto_config.utils` while the definitions live in one place.
+Platform = _platform_module.Platform
+PlatformEnum = _platform_module.PlatformEnum
+CpuArchEnum = _platform_module.CpuArchEnum
+current_platform = _platform_module.current_platform
+detect_hardware = _platform_module.detect_hardware
 
 
 def check_device_count(
@@ -410,7 +117,7 @@ def check_device_count(
     return True
 
 
-class BaseTuner(ABC):
+class BaseTuner:
     """Base class for per-model auto-tuning profiles.
 
     A profile subclasses this, sets `MODEL_TYPE`, and implements the two
